cli: move filter utils to filters.go and removed errors.go file
Before, the utils.go was cluttered with filter functions, so that's moved to filters.go. On the other hand, errors.go only had a single function, so that was moved to utils.go
